Move PortfolioStockCard to its owm composable
