+++ app/src/main/java/com/farmingdale/stockscreener/views/home/Portfolio.kt
@@ -3,10 +3,8 @@ package com.farmingdale.stockscreener.views.home
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Card
@@ -16,14 +14,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.farmingdale.stockscreener.R
 import com.farmingdale.stockscreener.model.local.SimpleQuoteData
-import com.farmingdale.stockscreener.ui.theme.indexColor
-import com.farmingdale.stockscreener.ui.theme.negativeTextColor
-import com.farmingdale.stockscreener.ui.theme.positiveTextColor
+import com.farmingdale.stockscreener.views.stock.StockCard
 
 @Composable
 fun Portfolio(
@@ -65,70 +59,10 @@ fun Portfolio(
                         items = watchList,
                         key = { quote -> quote.symbol }
                     ) { quote ->
-                        PortfolioStockCard(quote = quote)
+                        StockCard(quote = quote)
                     }
                 }
             )
         }
-    }
-}
-
-@Composable
-fun PortfolioStockCard(quote: SimpleQuoteData) {
-    Card(
-        modifier = Modifier
-            .size(125.dp, 75.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(8.dp),
-            verticalArrangement = Arrangement.SpaceAround
-        ) {
-            Text(
-                text = quote.name,
-                style = MaterialTheme.typography.titleSmall,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                color = indexColor
-            )
-            Text(
-                text = quote.price.toString(),
-                style = MaterialTheme.typography.titleSmall
-            )
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween
-            ) {
-                Text(
-                    text = quote.change,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = if (quote.change.contains('+')) positiveTextColor else negativeTextColor
-                )
-                Text(
-                    text = quote.percentChange,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = if (quote.change.contains('+')) positiveTextColor else negativeTextColor
-                )
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun PreviewPortfolioStockCard() {
-    Row {
-        val dummyQuote = SimpleQuoteData(
-            symbol = "DUMMY",
-            name = "Dummy Inc.",
-            price = 100.0,
-            change = "+10.0",
-            percentChange = "+10%"
-        )
-        PortfolioStockCard(
-            quote = dummyQuote
-        )
-        PortfolioStockCard(quote = dummyQuote)
     }
 }
